feat: touch head method, and refact get, utils ~
